refactor: extract register API into routes_api.py blueprint
Move /api/register out of app.py into a new routes_api.py blueprint
named "api". Behavior unchanged.

State previously held as module-level globals in app.py and used
only by this handler moves with it:
- unauthorized_log_tracker — per-IP rate limit for unauthorized
  access WARNING log lines
- failed_icon_cache / RETRY_INTERVAL — pass-through to
  image_utils.resolve_image_metadata to throttle icon-download
  retries on the upstream CDN

Captured references to `app.*` become `current_app.*`:
- app.config['api_token'] -> current_app.config['api_token']
- app.debug -> current_app.debug
- IMAGE_DIR -> current_app.config['IMAGE_DIR'] (set in Step 5)

Drops now-unused imports from app.py: request, jsonify,
resolve_image_metadata, parse_bool, and the previously-unused
flask_sqlalchemy.SQLAlchemy alias (db comes from extensions).

Phase 5 (a later phase, not this one) will add the canonical
/api/v1/register endpoint in this same module.
